refactor(challenges): inline single-use helpers, drop over-factored .utils files
- Inline shouldShowPendingReviewBadge directly into ImagesCard (4-line owner+REVIEW
  predicate didn't warrant its own module + micro-test).
- Move selectPayableUsers back into challenge-rewards.ts as a module-local function
  and drop its unit test (the .utils split existed only to dodge the prisma-client
  import crash under vitest).
- Keep the parseChallengeMetadata reconciliation test (db-free, still valuable).
