+++ src/components/Image/Infinite/ImagesCard.tsx
@@ -6,7 +6,6 @@ import HoverActionButton from '~/components/Cards/components/HoverActionButton';
 import { RoutedDialogLink } from '~/components/Dialog/RoutedDialogLink';
 import { DurationBadge } from '~/components/DurationBadge/DurationBadge';
 import { JudgeScoreBadge } from '~/components/Image/JudgeScoreBadge/JudgeScoreBadge';
-import { shouldShowPendingReviewBadge } from '~/components/Image/Infinite/pending-review-badge.utils';
 import { EdgeMedia2 } from '~/components/EdgeMedia/EdgeMedia';
 import { getSkipValue } from '~/components/EdgeMedia/EdgeMedia.util';
 import { ImageContextMenu } from '~/components/Image/ContextMenu/ImageContextMenu';
@@ -24,7 +23,7 @@ import { VotableTags } from '~/components/VotableTags/VotableTags';
 import { useFeatureFlags } from '~/providers/FeatureFlagsProvider';
 import type { ImagesInfiniteModel } from '~/server/services/image.service';
 import { getIsPublicBrowsingLevel } from '~/shared/constants/browsingLevel.constants';
-import { ImageIngestionStatus, MediaType } from '~/shared/utils/prisma/enums';
+import { CollectionItemStatus, ImageIngestionStatus, MediaType } from '~/shared/utils/prisma/enums';
 import { generationGraphPanel } from '~/store/generation-graph.store';
 import { useImageStore } from '~/store/image.store';
 import { useTourContext } from '~/components/Tours/ToursProvider';
@@ -169,7 +168,8 @@ function ImagesCardContent({ data, height }: { data: ImagesInfiniteModel; height
                         POI
                       </Badge>
                     )}
-                    {shouldShowPendingReviewBadge(data, currentUser?.id) && (
+                    {currentUser?.id === data.userId &&
+                      data.collectionItemStatus === CollectionItemStatus.REVIEW && (
                       <Tooltip label="Still being reviewed — not yet eligible for judging" withinPortal>
                         <Badge variant="filled" radius="xl" h={26} color="yellow">
                           Pending review
