--- src/components/Image/Infinite/pending-review-badge.utils.ts
@@ -1,12 +1,0 @@
-import { CollectionItemStatus } from '~/shared/utils/prisma/enums';
-
-export function shouldShowPendingReviewBadge(
-  image: { userId: number; collectionItemStatus?: string | null },
-  currentUserId?: number
-): boolean {
-  return (
-    !!currentUserId &&
-    image.userId === currentUserId &&
-    image.collectionItemStatus === CollectionItemStatus.REVIEW
-  );
-}
