+++ src/server/games/daily-challenge/__tests__/challenge-rewards.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect } from 'vitest';
 import { parseChallengeMetadata } from '~/server/schema/challenge.schema';
-import { selectPayableUsers } from '~/server/games/daily-challenge/challenge-rewards.utils';
 
 describe('parseChallengeMetadata reconciliation', () => {
   it('round-trips the reconciliation field', () => {
@@ -13,17 +12,5 @@ describe('parseChallengeMetadata reconciliation', () => {
 
   it('defaults reconciliation to undefined when absent', () => {
     expect(parseChallengeMetadata({ themeElements: ['a'] }).reconciliation).toBeUndefined();
-  });
-});
-
-describe('selectPayableUsers', () => {
-  it('removes excluded users (winners ∪ already-paid)', () => {
-    expect(selectPayableUsers([1, 2, 3, 4], [2, 4])).toEqual([1, 3]);
-  });
-  it('returns empty when all excluded', () => {
-    expect(selectPayableUsers([1, 2], [1, 2, 3])).toEqual([]);
-  });
-  it('de-duplicates qualifiers', () => {
-    expect(selectPayableUsers([1, 1, 2], [])).toEqual([1, 2]);
   });
 });
